Split into a slim core plus one file per tool, with a real name-lookup dispatch registry
The registry/dispatcher/consent-gate/round-gate machinery moves into its
own file, and each tool gets its own file under Features/Chat/Tools/.
Dispatch goes through name-keyed lookup tables instead of hardcoded
switch statements, so adding a tool is a registration, not a new case in
two separate switches.
