+++ Sources/Nativ/Features/Chat/ChatView.swift
@@ -12,741 +12,6 @@ struct ChatQueuedPrompt: Identifiable, Equatable {
     let position: Int
 }
 
-private enum ChatImageToolRegistry {
-    static func definitions(canEdit: Bool) -> [MLXChatToolDefinition] {
-        var tools = [tool(
-            name: "generate_image",
-            description: "Create one or more new images from a detailed text prompt."
-        )]
-        if canEdit {
-            tools.append(tool(
-                name: "edit_image",
-                description: "Edit the most recently attached or generated image using a text instruction."
-            ))
-        }
-        return tools
-    }
-
-    private static func tool(name: String, description: String) -> MLXChatToolDefinition {
-        MLXChatToolDefinition(function: MLXChatFunctionDefinition(
-            name: name,
-            description: description,
-            parameters: .object([
-                "type": .string("object"),
-                "additionalProperties": .bool(false),
-                "properties": .object([
-                    "prompt": .object([
-                        "type": .string("string"),
-                        "description": .string("A specific visual description or edit instruction.")
-                    ]),
-                    "width": .object([
-                        "type": .string("integer"),
-                        "minimum": .number(64),
-                        "maximum": .number(2048)
-                    ]),
-                    "height": .object([
-                        "type": .string("integer"),
-                        "minimum": .number(64),
-                        "maximum": .number(2048)
-                    ]),
-                    "count": .object([
-                        "type": .string("integer"),
-                        "minimum": .number(1),
-                        "maximum": .number(4)
-                    ]),
-                    "seed": .object([
-                        "type": .array([.string("integer"), .string("null")])
-                    ])
-                ]),
-                "required": .array([.string("prompt")])
-            ])
-        ))
-    }
-}
-
-private struct ChatImageToolArguments: Decodable {
-    let prompt: String
-    let width: Int?
-    let height: Int?
-    let count: Int?
-    let seed: Int?
-}
-
-private struct ChatImageToolResultPayload: Encodable {
-    struct Image: Encodable {
-        let attachmentID: String
-        let width: Int
-        let height: Int
-        let seed: Int
-
-        enum CodingKeys: String, CodingKey {
-            case attachmentID = "attachment_id"
-            case width
-            case height
-            case seed
-        }
-    }
-
-    let ok: Bool
-    let operation: String
-    let images: [Image]?
-    let error: String?
-}
-
-private struct ChatImageToolExecution {
-    let content: String
-    let attachments: [ChatImageAttachment]
-}
-
-private enum ChatImageToolError: LocalizedError {
-    case unsupportedTool(String)
-    case invalidArguments
-    case emptyPrompt
-    case missingReference
-
-    var errorDescription: String? {
-        switch self {
-        case .unsupportedTool(let name):
-            "Unsupported image tool: \(name)"
-        case .invalidArguments:
-            "The image tool arguments were not valid JSON."
-        case .emptyPrompt:
-            "The image prompt cannot be empty."
-        case .missingReference:
-            "No earlier image is available to edit."
-        }
-    }
-}
-
-private struct ChatImageToolExecutor {
-    func execute(
-        call: MLXChatToolCall,
-        modelID: String,
-        baseURL: URL,
-        apiKey: String?,
-        references: [ChatImageAttachment]
-    ) async throws -> ChatImageToolExecution {
-        guard let name = call.function?.name else {
-            throw ChatImageToolError.unsupportedTool("unknown")
-        }
-        guard name == "generate_image" || name == "edit_image" else {
-            throw ChatImageToolError.unsupportedTool(name)
-        }
-        guard let arguments = call.function?.arguments?.data(using: .utf8),
-              let decoded = try? JSONDecoder().decode(ChatImageToolArguments.self, from: arguments)
-        else {
-            throw ChatImageToolError.invalidArguments
-        }
-
-        let prompt = decoded.prompt.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !prompt.isEmpty else {
-            throw ChatImageToolError.emptyPrompt
-        }
-        if name == "edit_image", references.isEmpty {
-            throw ChatImageToolError.missingReference
-        }
-
-        let sourceSize = name == "edit_image" ? imageSize(for: references.first) : nil
-        let settings = ImageRequestSettings(
-            count: min(max(decoded.count ?? 1, 1), 4),
-            width: boundedDimension(decoded.width ?? sourceSize?.width ?? 512),
-            height: boundedDimension(decoded.height ?? sourceSize?.height ?? 512),
-            steps: 4,
-            guidance: 1,
-            seedText: decoded.seed.map(String.init) ?? ""
-        )
-        let outputs = try await ImageGenerationExecutor().run(
-            baseURL: baseURL,
-            apiKey: apiKey,
-            modelID: modelID,
-            prompt: prompt,
-            references: name == "edit_image" ? references : [],
-            settings: settings,
-            seed: decoded.seed
-        )
-        let attachments = outputs.map(\.attachment)
-        let payload = ChatImageToolResultPayload(
-            ok: true,
-            operation: name == "edit_image" ? "edit" : "generate",
-            images: outputs.map {
-                ChatImageToolResultPayload.Image(
-                    attachmentID: $0.id.uuidString,
-                    width: $0.width,
-                    height: $0.height,
-                    seed: $0.seed
-                )
-            },
-            error: nil
-        )
-        return ChatImageToolExecution(
-            content: try encodedPayload(payload),
-            attachments: attachments
-        )
-    }
-
-    func failurePayload(operation: String, error: Error) -> String {
-        let payload = ChatImageToolResultPayload(
-            ok: false,
-            operation: operation,
-            images: nil,
-            error: error.localizedDescription
-        )
-        return (try? encodedPayload(payload))
-            ?? #"{"ok":false,"error":"Image tool failed."}"#
-    }
-
-    private func boundedDimension(_ value: Int) -> Int {
-        min(max((value / 16) * 16, 64), 2_048)
-    }
-
-    private func imageSize(for attachment: ChatImageAttachment?) -> ImageGenerationPixelSize? {
-        guard let data = attachment?.imageData,
-              let image = NSImage(data: data),
-              image.size.width > 0,
-              image.size.height > 0
-        else {
-            return nil
-        }
-        return ImageGenerationPixelSize(
-            width: Int(image.size.width.rounded()),
-            height: Int(image.size.height.rounded())
-        )
-    }
-
-    private func encodedPayload(_ payload: ChatImageToolResultPayload) throws -> String {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = [.sortedKeys]
-        return String(decoding: try encoder.encode(payload), as: UTF8.self)
-    }
-}
-
-private enum ChatSystemMonitorToolRegistry {
-    static let toolName = "get_system_stats"
-
-    static func definitions() -> [MLXChatToolDefinition] {
-        [MLXChatToolDefinition(function: MLXChatFunctionDefinition(
-            name: toolName,
-            description: "Get current CPU, GPU, memory, and disk usage on this Mac.",
-            parameters: .object([
-                "type": .string("object"),
-                "additionalProperties": .bool(false),
-                "properties": .object([:])
-            ])
-        ))]
-    }
-}
-
-private struct ChatSystemMonitorToolResultPayload: Encodable {
-    let ok: Bool
-    let cpuUsagePercent: Int?
-    let gpuUsagePercent: Int?
-    let memoryUsedGB: Double?
-    let memoryTotalGB: Double?
-    let diskUsedGB: Double?
-    let diskTotalGB: Double?
-    let uptimeSeconds: Int?
-    let error: String?
-
-    enum CodingKeys: String, CodingKey {
-        case ok
-        case cpuUsagePercent = "cpu_usage_percent"
-        case gpuUsagePercent = "gpu_usage_percent"
-        case memoryUsedGB = "memory_used_gb"
-        case memoryTotalGB = "memory_total_gb"
-        case diskUsedGB = "disk_used_gb"
-        case diskTotalGB = "disk_total_gb"
-        case uptimeSeconds = "uptime_seconds"
-        case error
-    }
-}
-
-private struct ChatSystemMonitorToolExecutor {
-    @MainActor
-    func execute(call: MLXChatToolCall) async throws -> String {
-        guard call.function?.name == ChatSystemMonitorToolRegistry.toolName else {
-            throw ChatImageToolError.unsupportedTool(call.function?.name ?? "unknown")
-        }
-
-        let store = SystemMonitorStore()
-        // First read only seeds the previous-tick baseline; second read
-        // after a delay gives a real CPU/disk delta.
-        _ = await store.collectSnapshot()
-        try await Task.sleep(nanoseconds: 300_000_000)
-        let snapshot = await store.collectSnapshot()
-
-        let payload = ChatSystemMonitorToolResultPayload(
-            ok: true,
-            cpuUsagePercent: percent(snapshot.cpu.totalUsage),
-            gpuUsagePercent: snapshot.gpu.deviceUsage.flatMap(percent),
-            memoryUsedGB: gigabytes(snapshot.memory.usedBytes),
-            memoryTotalGB: gigabytes(snapshot.memory.totalBytes),
-            diskUsedGB: gigabytes(snapshot.disk.usedBytes),
-            diskTotalGB: gigabytes(snapshot.disk.totalBytes),
-            uptimeSeconds: Int(snapshot.uptime),
-            error: nil
-        )
-        return try encodedPayload(payload)
-    }
-
-    func failurePayload(operation: String, error: Error) -> String {
-        let payload = ChatSystemMonitorToolResultPayload(
-            ok: false,
-            cpuUsagePercent: nil,
-            gpuUsagePercent: nil,
-            memoryUsedGB: nil,
-            memoryTotalGB: nil,
-            diskUsedGB: nil,
-            diskTotalGB: nil,
-            uptimeSeconds: nil,
-            error: error.localizedDescription
-        )
-        return (try? encodedPayload(payload))
-            ?? #"{"ok":false,"error":"System monitor tool failed."}"#
-    }
-
-    private func percent(_ usage: Double) -> Int {
-        Int((usage * 100).rounded())
-    }
-
-    private func gigabytes(_ bytes: UInt64) -> Double {
-        (Double(bytes) / 1_073_741_824 * 100).rounded() / 100
-    }
-
-    private func encodedPayload(_ payload: ChatSystemMonitorToolResultPayload) throws -> String {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = [.sortedKeys]
-        return String(decoding: try encoder.encode(payload), as: UTF8.self)
-    }
-}
-
-private enum ChatModelLibraryToolRegistry {
-    static let toolName = "list_models"
-
-    static func definitions() -> [MLXChatToolDefinition] {
-        [MLXChatToolDefinition(function: MLXChatFunctionDefinition(
-            name: toolName,
-            description: "List the MLX models already downloaded on this Mac, with size and quantization.",
-            parameters: .object([
-                "type": .string("object"),
-                "additionalProperties": .bool(false),
-                "properties": .object([:])
-            ])
-        ))]
-    }
-}
-
-private struct ChatModelLibraryToolResultPayload: Encodable {
-    struct Model: Encodable {
-        let repoID: String
-        let sizeGB: Double?
-        let parameterCount: Int64?
-        let quantizationBits: Int?
-        let capabilities: [String]
-
-        enum CodingKeys: String, CodingKey {
-            case repoID = "repo_id"
-            case sizeGB = "size_gb"
-            case parameterCount = "parameter_count"
-            case quantizationBits = "quantization_bits"
-            case capabilities
-        }
-    }
-
-    let ok: Bool
-    let models: [Model]?
-    let error: String?
-}
-
-private struct ChatModelLibraryToolExecutor {
-    func execute(call: MLXChatToolCall, context: ChatToolExecutionContext) async throws -> String {
-        guard call.function?.name == ChatModelLibraryToolRegistry.toolName else {
-            throw ChatImageToolError.unsupportedTool(call.function?.name ?? "unknown")
-        }
-
-        let models = try await LocalModelDiscovery.scan(
-            path: context.modelSearchPath,
-            additionalPaths: context.additionalModelSearchPaths
-        )
-        let payload = ChatModelLibraryToolResultPayload(
-            ok: true,
-            models: models.map { model in
-                ChatModelLibraryToolResultPayload.Model(
-                    repoID: model.repoID,
-                    sizeGB: model.sizeBytes.map(gigabytes),
-                    parameterCount: model.parameterCount,
-                    quantizationBits: model.quantizationBits,
-                    capabilities: model.capabilities.map(\.rawValue).sorted()
-                )
-            },
-            error: nil
-        )
-        return try encodedPayload(payload)
-    }
-
-    func failurePayload(operation: String, error: Error) -> String {
-        let payload = ChatModelLibraryToolResultPayload(ok: false, models: nil, error: error.localizedDescription)
-        return (try? encodedPayload(payload))
-            ?? #"{"ok":false,"error":"Model library tool failed."}"#
-    }
-
-    private func gigabytes(_ bytes: Int64) -> Double {
-        (Double(bytes) / 1_073_741_824 * 100).rounded() / 100
-    }
-
-    private func encodedPayload(_ payload: ChatModelLibraryToolResultPayload) throws -> String {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = [.sortedKeys]
-        return String(decoding: try encoder.encode(payload), as: UTF8.self)
-    }
-}
-
-private enum ChatServerStatsToolRegistry {
-    static let toolName = "get_server_stats"
-
-    static func definitions() -> [MLXChatToolDefinition] {
-        [MLXChatToolDefinition(function: MLXChatFunctionDefinition(
-            name: toolName,
-            description: "Get this Mac's local model server performance stats: requests, tokens, speed, time to first token.",
-            parameters: .object([
-                "type": .string("object"),
-                "additionalProperties": .bool(false),
-                "properties": .object([:])
-            ])
-        ))]
-    }
-}
-
-private struct ChatServerStatsToolResultPayload: Encodable {
-    let ok: Bool
-    let requestsCompleted: Int?
-    let requestsFailed: Int?
-    let promptTokensTotal: Int?
-    let completionTokensTotal: Int?
-    let generatedTokensTotal: Int?
-    let avgDecodeTokensPerSecond: Double?
-    let avgRequestTokensPerSecond: Double?
-    let avgTimeToFirstTokenMs: Double?
-    let peakMemoryGB: Double?
-    let error: String?
-
-    enum CodingKeys: String, CodingKey {
-        case ok
-        case requestsCompleted = "requests_completed"
-        case requestsFailed = "requests_failed"
-        case promptTokensTotal = "prompt_tokens_total"
-        case completionTokensTotal = "completion_tokens_total"
-        case generatedTokensTotal = "generated_tokens_total"
-        case avgDecodeTokensPerSecond = "avg_decode_tokens_per_second"
-        case avgRequestTokensPerSecond = "avg_request_tokens_per_second"
-        case avgTimeToFirstTokenMs = "avg_time_to_first_token_ms"
-        case peakMemoryGB = "peak_memory_gb"
-        case error
-    }
-}
-
-private struct ChatServerStatsToolExecutor {
-    func execute(call: MLXChatToolCall) throws -> String {
-        guard call.function?.name == ChatServerStatsToolRegistry.toolName else {
-            throw ChatImageToolError.unsupportedTool(call.function?.name ?? "unknown")
-        }
-
-        let summary = NativAnalyticsStore().fetchSummary()
-        let payload = ChatServerStatsToolResultPayload(
-            ok: true,
-            requestsCompleted: summary.requestsCompleted,
-            requestsFailed: summary.requestsFailed,
-            promptTokensTotal: summary.promptTokensTotal,
-            completionTokensTotal: summary.completionTokensTotal,
-            generatedTokensTotal: summary.generatedTokensTotal,
-            avgDecodeTokensPerSecond: rounded(summary.averageDecodeTokensPerSecond),
-            avgRequestTokensPerSecond: rounded(summary.averageRequestTokensPerSecond),
-            avgTimeToFirstTokenMs: rounded(summary.averageTTFTMilliseconds),
-            peakMemoryGB: summary.peakMemoryBytesMax.map(gigabytes),
-            error: nil
-        )
-        return try encodedPayload(payload)
-    }
-
-    func failurePayload(operation: String, error: Error) -> String {
-        let payload = ChatServerStatsToolResultPayload(
-            ok: false,
-            requestsCompleted: nil,
-            requestsFailed: nil,
-            promptTokensTotal: nil,
-            completionTokensTotal: nil,
-            generatedTokensTotal: nil,
-            avgDecodeTokensPerSecond: nil,
-            avgRequestTokensPerSecond: nil,
-            avgTimeToFirstTokenMs: nil,
-            peakMemoryGB: nil,
-            error: error.localizedDescription
-        )
-        return (try? encodedPayload(payload))
-            ?? #"{"ok":false,"error":"Server stats tool failed."}"#
-    }
-
-    private func rounded(_ value: Double?) -> Double? {
-        value.map { ($0 * 100).rounded() / 100 }
-    }
-
-    private func gigabytes(_ bytes: Int64) -> Double {
-        (Double(bytes) / 1_073_741_824 * 100).rounded() / 100
-    }
-
-    private func encodedPayload(_ payload: ChatServerStatsToolResultPayload) throws -> String {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = [.sortedKeys]
-        return String(decoding: try encoder.encode(payload), as: UTF8.self)
-    }
-}
-
-private enum ChatSwitchModelToolRegistry {
-    static let toolName = "switch_model"
-
-    static func definitions() -> [MLXChatToolDefinition] {
-        [MLXChatToolDefinition(function: MLXChatFunctionDefinition(
-            name: toolName,
-            description: "Switch the active local language model. Requires user confirmation before it runs.",
-            parameters: .object([
-                "type": .string("object"),
-                "additionalProperties": .bool(false),
-                "properties": .object([
-                    "model_id": .object([
-                        "type": .string("string"),
-                        "description": .string("Exact model identifier to switch to, e.g. mlx-community/Qwen3-1.7B-4bit")
-                    ])
-                ]),
-                "required": .array([.string("model_id")])
-            ])
-        ))]
-    }
-}
-
-private struct ChatSwitchModelToolArguments: Decodable {
-    let modelID: String
-
-    enum CodingKeys: String, CodingKey {
-        case modelID = "model_id"
-    }
-}
-
-private struct ChatSwitchModelToolResultPayload: Encodable {
-    let ok: Bool
-    let previousModelID: String?
-    let newModelID: String?
-    let changed: Bool?
-    let declined: Bool?
-    let error: String?
-
-    enum CodingKeys: String, CodingKey {
-        case ok
-        case previousModelID = "previous_model_id"
-        case newModelID = "new_model_id"
-        case changed
-        case declined
-        case error
-    }
-}
-
-private enum ChatSwitchModelToolError: LocalizedError {
-    case invalidArguments
-    case timedOut
-    case switchFailed
-    case mismatchedModel(requested: String, active: String?)
-
-    var errorDescription: String? {
-        switch self {
-        case .invalidArguments:
-            return "The switch_model arguments were not valid JSON."
-        case .timedOut:
-            return "The model switch did not finish in time."
-        case .switchFailed:
-            return "The server failed to start with the requested model."
-        case .mismatchedModel(let requested, let active):
-            return "Requested \(requested) but the active model is now \(active ?? "unknown")."
-        }
-    }
-}
-
-private struct ChatSwitchModelToolExecutor {
-    @MainActor
-    func execute(call: MLXChatToolCall, appModel: NativModel) async throws -> String {
-        guard call.function?.name == ChatSwitchModelToolRegistry.toolName else {
-            throw ChatImageToolError.unsupportedTool(call.function?.name ?? "unknown")
-        }
-        guard let argumentsData = call.function?.arguments?.data(using: .utf8),
-              let arguments = try? JSONDecoder().decode(ChatSwitchModelToolArguments.self, from: argumentsData)
-        else {
-            throw ChatSwitchModelToolError.invalidArguments
-        }
-
-        let previousModelID = appModel.settings.normalized().languageModelID
-        if previousModelID == arguments.modelID {
-            return try encodedPayload(ChatSwitchModelToolResultPayload(
-                ok: true,
-                previousModelID: previousModelID,
-                newModelID: arguments.modelID,
-                changed: false,
-                declined: false,
-                error: nil
-            ))
-        }
-
-        appModel.switchLanguageModel(to: arguments.modelID)
-
-        let deadline = Date().addingTimeInterval(60)
-        while appModel.modelSwitchInProgress, Date() < deadline {
-            try await Task.sleep(nanoseconds: 200_000_000)
-        }
-        guard !appModel.modelSwitchInProgress else {
-            throw ChatSwitchModelToolError.timedOut
-        }
-        guard appModel.isRunning else {
-            throw ChatSwitchModelToolError.switchFailed
-        }
-
-        let activeModelID = appModel.settings.normalized().languageModelID
-        guard activeModelID == arguments.modelID else {
-            throw ChatSwitchModelToolError.mismatchedModel(
-                requested: arguments.modelID,
-                active: activeModelID
-            )
-        }
-
-        return try encodedPayload(ChatSwitchModelToolResultPayload(
-            ok: true,
-            previousModelID: previousModelID,
-            newModelID: activeModelID,
-            changed: true,
-            declined: false,
-            error: nil
-        ))
-    }
-
-    func declinedPayload() -> String {
-        (try? encodedPayload(ChatSwitchModelToolResultPayload(
-            ok: false,
-            previousModelID: nil,
-            newModelID: nil,
-            changed: nil,
-            declined: true,
-            error: "The user declined this model switch."
-        ))) ?? #"{"ok":false,"declined":true,"error":"The user declined this model switch."}"#
-    }
-
-    func failurePayload(operation: String, error: Error) -> String {
-        let payload = ChatSwitchModelToolResultPayload(
-            ok: false,
-            previousModelID: nil,
-            newModelID: nil,
-            changed: nil,
-            declined: false,
-            error: error.localizedDescription
-        )
-        return (try? encodedPayload(payload))
-            ?? #"{"ok":false,"error":"Model switch failed."}"#
-    }
-
-    private func encodedPayload(_ payload: ChatSwitchModelToolResultPayload) throws -> String {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = [.sortedKeys]
-        return String(decoding: try encoder.encode(payload), as: UTF8.self)
-    }
-}
-
-private struct ChatToolExecutionContext {
-    let imageGenerationModelID: String?
-    let baseURL: URL
-    let apiKey: String?
-    let imageReferences: [ChatImageAttachment]
-    let modelSearchPath: String
-    let additionalModelSearchPaths: [String]
-}
-
-private struct ChatToolExecutionOutcome {
-    let content: String
-    let attachments: [ChatImageAttachment]
-}
-
-private enum ChatToolRegistry {
-    static func definitions(
-        context: ChatToolExecutionContext,
-        canEditImage: Bool
-    ) -> [MLXChatToolDefinition] {
-        var tools: [MLXChatToolDefinition] = []
-        if context.imageGenerationModelID?.isEmpty == false {
-            tools.append(contentsOf: ChatImageToolRegistry.definitions(canEdit: canEditImage))
-        }
-        tools.append(contentsOf: ChatSystemMonitorToolRegistry.definitions())
-        tools.append(contentsOf: ChatModelLibraryToolRegistry.definitions())
-        tools.append(contentsOf: ChatServerStatsToolRegistry.definitions())
-        tools.append(contentsOf: ChatSwitchModelToolRegistry.definitions())
-        return tools
-    }
-}
-
-private enum ChatToolDispatcher {
-    static func execute(
-        call: MLXChatToolCall,
-        context: ChatToolExecutionContext
-    ) async throws -> ChatToolExecutionOutcome {
-        switch call.function?.name {
-        case "generate_image", "edit_image":
-            guard let imageModelID = context.imageGenerationModelID else {
-                throw ChatImageToolError.unsupportedTool(call.function?.name ?? "image")
-            }
-            let result = try await ChatImageToolExecutor().execute(
-                call: call,
-                modelID: imageModelID,
-                baseURL: context.baseURL,
-                apiKey: context.apiKey,
-                references: context.imageReferences
-            )
-            return ChatToolExecutionOutcome(content: result.content, attachments: result.attachments)
-        case ChatSystemMonitorToolRegistry.toolName:
-            let content = try await ChatSystemMonitorToolExecutor().execute(call: call)
-            return ChatToolExecutionOutcome(content: content, attachments: [])
-        case ChatModelLibraryToolRegistry.toolName:
-            let content = try await ChatModelLibraryToolExecutor().execute(call: call, context: context)
-            return ChatToolExecutionOutcome(content: content, attachments: [])
-        case ChatServerStatsToolRegistry.toolName:
-            let content = try ChatServerStatsToolExecutor().execute(call: call)
-            return ChatToolExecutionOutcome(content: content, attachments: [])
-        default:
-            throw ChatImageToolError.unsupportedTool(call.function?.name ?? "unknown")
-        }
-    }
-
-    static func failurePayload(toolName: String?, error: Error) -> String {
-        switch toolName {
-        case "generate_image", "edit_image":
-            return ChatImageToolExecutor().failurePayload(operation: toolName ?? "image", error: error)
-        case ChatSystemMonitorToolRegistry.toolName:
-            return ChatSystemMonitorToolExecutor().failurePayload(
-                operation: toolName ?? ChatSystemMonitorToolRegistry.toolName,
-                error: error
-            )
-        case ChatModelLibraryToolRegistry.toolName:
-            return ChatModelLibraryToolExecutor().failurePayload(
-                operation: toolName ?? ChatModelLibraryToolRegistry.toolName,
-                error: error
-            )
-        case ChatServerStatsToolRegistry.toolName:
-            return ChatServerStatsToolExecutor().failurePayload(
-                operation: toolName ?? ChatServerStatsToolRegistry.toolName,
-                error: error
-            )
-        case ChatSwitchModelToolRegistry.toolName:
-            return ChatSwitchModelToolExecutor().failurePayload(
-                operation: toolName ?? ChatSwitchModelToolRegistry.toolName,
-                error: error
-            )
-        default:
-            return ChatImageToolExecutor().failurePayload(operation: toolName ?? "tool", error: error)
-        }
-    }
-}
-
 struct ChatView: View {
     private enum Layout {
         static let conversationMaxWidth: CGFloat = 680
@@ -940,7 +205,7 @@ final class ChatViewModel: ObservableObject {
     private var streamFlushDates: [UUID: Date] = [:]
     private var streamFlushTasks: [UUID: Task<Void, Never>] = [:]
     private weak var appModel: NativModel?
-    private var pendingToolConsent: [UUID: CheckedContinuation<Bool, Never>] = [:]
+    private let toolConsentGate = ChatToolConsentGate()
 
     init() {
         storedSessions = sessionStore.loadSessions()
@@ -1180,23 +445,15 @@ final class ChatViewModel: ObservableObject {
     }
 
     func confirmToolConsent(_ toolMessageID: UUID) {
-        pendingToolConsent.removeValue(forKey: toolMessageID)?.resume(returning: true)
+        toolConsentGate.confirm(toolMessageID)
     }
 
     func denyToolConsent(_ toolMessageID: UUID) {
-        pendingToolConsent.removeValue(forKey: toolMessageID)?.resume(returning: false)
+        toolConsentGate.deny(toolMessageID)
     }
 
     private func awaitToolConsent(for toolMessageID: UUID) async -> Bool {
-        await withTaskCancellationHandler {
-            await withCheckedContinuation { (continuation: CheckedContinuation<Bool, Never>) in
-                pendingToolConsent[toolMessageID] = continuation
-            }
-        } onCancel: { [weak self] in
-            Task { @MainActor in
-                self?.pendingToolConsent.removeValue(forKey: toolMessageID)?.resume(returning: false)
-            }
-        }
+        await toolConsentGate.awaitDecision(for: toolMessageID)
     }
 
     func cancel() {
@@ -1373,11 +630,10 @@ final class ChatViewModel: ObservableObject {
         )
         var assistantMessageID = queuedRequest.assistantMessageID
         var toolRounds = 0
-        let maximumToolRounds = 4
 
         while true {
             try Task.checkCancellation()
-            let advertisesTools = toolRounds < maximumToolRounds
+            let advertisesTools = ChatToolRoundGate.advertisesTools(atRound: toolRounds)
             guard let request = makeCompletionRequest(
                 for: queuedRequest,
                 before: assistantMessageID,
@@ -1433,9 +689,8 @@ final class ChatViewModel: ObservableObject {
                         attachments: []
                     )
                     let approved = await awaitToolConsent(for: toolMessageID)
-                    do {
-                        try Task.checkCancellation()
-                    } catch {
+                    switch ChatToolConsentRouter.outcome(approved: approved, isCancelled: Task.isCancelled) {
+                    case .cancelled:
                         cancelToolMessages(
                             currentID: toolMessageID,
                             currentCall: toolCall,
@@ -1444,9 +699,7 @@ final class ChatViewModel: ObservableObject {
                             in: queuedRequest.sessionID
                         )
                         throw CancellationError()
-                    }
-
-                    guard approved else {
+                    case .declined:
                         updateToolMessage(
                             toolMessageID,
                             in: queuedRequest.sessionID,
@@ -1455,6 +708,8 @@ final class ChatViewModel: ObservableObject {
                             attachments: []
                         )
                         continue
+                    case .approved:
+                        break
                     }
                     guard let appModel else {
                         updateToolMessage(
