+++ project.yml
@@ -102,6 +102,22 @@ targets:
       - path: Sources/Nativ/Features/Integrations/IntegrationServices.swift
       - path: Sources/Nativ/Features/Integrations/IntegrationTypes.swift
       - path: Sources/Nativ/Utilities/ShellEnvironment.swift
+      - path: Sources/Nativ/Features/Chat/ChatToolRegistry.swift
+      - path: Sources/Nativ/Features/Chat/ChatImageTool.swift
+      - path: Sources/Nativ/Features/Chat/ChatSystemStatsTool.swift
+      - path: Sources/Nativ/Features/Chat/ChatListModelsTool.swift
+      - path: Sources/Nativ/Features/Chat/ChatServerStatsTool.swift
+      - path: Sources/Nativ/Features/Chat/ChatSwitchModelTool.swift
+      - path: Sources/Nativ/Features/Chat/ChatSessionStore.swift
+      - path: Sources/Nativ/Features/ImageGeneration/ImageGenerationViewModel.swift
+      - path: Sources/Nativ/Features/SystemMonitor/SystemMonitorStore.swift
+      - path: Sources/Nativ/Features/Models/LocalModelDiscovery.swift
+      - path: Sources/Nativ/Features/Dashboard/NativAnalyticsStore.swift
+      - path: Sources/Nativ/NativModel.swift
+      - path: Sources/Nativ/Features/Models/LocalModelProvider.swift
+      - path: Sources/Nativ/Utilities/Formatting.swift
+      - path: Sources/Nativ/Features/Models/HuggingFaceHub.swift
+      - path: Sources/Nativ/Features/Chat/ChatScreenCapture.swift
     dependencies:
       - target: NativServerKit
     settings:
